made anilist entity
