+++ japi/src/main/java/com/jamdb/japi/entities/watchList/Status.java
@@ -1,0 +1,6 @@
+package com.jamdb.japi.entities.watchList;
+
+public enum Status {
+    COMPLETED, DROPPED, PAUSED, PLAN_TO_WATCH, REWATCHING, WATCHING
+
+}
