+++ japi/src/main/java/com/jamdb/japi/services/UserService/UserService.java
@@ -3,7 +3,6 @@ package com.jamdb.japi.services.UserService;
 import com.jamdb.japi.dto.NewUserDto;
 import com.jamdb.japi.dto.UserAuthDto;
 import com.jamdb.japi.dto.UserResponse;
-import com.jamdb.japi.entities.content.Content;
 import com.jamdb.japi.entities.user.User;
 import com.jamdb.japi.entities.user.UserRole;
 import com.jamdb.japi.exceptions.UserAuthException;
@@ -17,10 +16,6 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.UUID;
-import java.util.stream.Collectors;
 
 @Service
 @Transactional
@@ -67,25 +62,25 @@ public class UserService implements UserServiceInterface {
         return UserResponse.builder().userName(user.getUsername()).email(user.getEmail()).token(jwtToken).build();
     }
 
-    @Override
-    public void addContent(String username, UUID content_id) {
-        var currentUser = userRepository.findByUserName(username);
-        contentService.findContent(content_id).ifPresent(content -> currentUser.ifPresent(user -> user.setAnime(content)));
-    }
+//    @Override
+//    public void addContent(String username, UUID content_id) {
+//        var currentUser = userRepository.findByUserName(username);
+//        contentService.findContent(content_id).ifPresent(content -> currentUser.ifPresent(user -> user.setAnime(content)));
+//    }
 
-    @Override
-    public List<Content> showContent(String username) throws UserAuthException {
-        var currentUser = userRepository.findByUserName(username).orElseThrow();
-        var contents = currentUser.getAnime();
-        return contents.stream().toList();
-    }
-
-    @Override
-    public void deleteContent(String username, UUID fromString) {
-        var user = userRepository.findByUserName(username).orElseThrow();
-        user.setAnimeList(user.getAnime().stream().filter(content -> !content.getId().equals(fromString)).collect(Collectors.toSet()));
-
-    }
+//    @Override
+//    public List<Content> showContent(String username) throws UserAuthException {
+//        var currentUser = userRepository.findByUserName(username).orElseThrow();
+//        var contents = currentUser.getAnime();
+//        return contents.stream().toList();
+//    }
+//
+//    @Override
+//    public void deleteContent(String username, UUID fromString) {
+//        var user = userRepository.findByUserName(username).orElseThrow();
+//        user.setAnimeList(user.getAnime().stream().filter(content -> !content.getId().equals(fromString)).collect(Collectors.toSet()));
+//
+//    }
 
 
 }
