+++ japi/src/main/java/com/jamdb/japi/services/UserService/UserServiceInterface.java
@@ -3,11 +3,7 @@ package com.jamdb.japi.services.UserService;
 import com.jamdb.japi.dto.NewUserDto;
 import com.jamdb.japi.dto.UserAuthDto;
 import com.jamdb.japi.dto.UserResponse;
-import com.jamdb.japi.entities.content.Content;
 import com.jamdb.japi.exceptions.UserAuthException;
-
-import java.util.List;
-import java.util.UUID;
 
 public interface UserServiceInterface {
     boolean userExists(String userName);
@@ -17,10 +13,10 @@ public interface UserServiceInterface {
     UserResponse getUser(String userName) throws UserAuthException;
 
     UserResponse authenticateUser(UserAuthDto userAuthDto) throws UserAuthException;
-
-    void addContent(String username, UUID content_id);
-
-    List<Content> showContent(String username) throws UserAuthException;
-
-    void deleteContent(String username, UUID fromString);
+//
+//    void addContent(String username, UUID content_id);
+//
+//    List<Content> showContent(String username) throws UserAuthException;
+//
+//    void deleteContent(String username, UUID fromString);
 }
